設定介面修改 解析度長寬 由字串改為Int

+++ car_password/app/src/main/java/com/example/carKotlinCode/jsonCommand.kt
@@ -75,8 +75,8 @@ class jsonCommand : Thread() {
     fun setQualityJSON(width: String, height: String) {
         val qualityJSONObject = JSONObject()
         qualityJSONObject.put("CMD", "SET_QUALITY")
-        qualityJSONObject.put("WIDTH", width)
-        qualityJSONObject.put("HEIGHT", height)
+        qualityJSONObject.put("WIDTH", width.toInt())
+        qualityJSONObject.put("HEIGHT", height.toInt())
         sendJsonToByteArray(qualityJSONObject)
     }
